INGM-276 Remove utils module
