+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 ## [Unreleased]
 ### Added
 - connect_servo_eoe_service and connect_servo_eoe_service_interface_index functions are added.
-- The utils module.
 
 ### Deprecated
 - connect_servo_ecat and connect_servo_ecat_interface_ip functions are removed.
