+++ tests/test_capture.py
@@ -2,9 +2,12 @@ import time
 
 import pytest
 import numpy as np
+from ingenialink.ethernet.register import EthernetRegister, REG_DTYPE, REG_ACCESS
+from ingenialink.canopen.register import CanopenRegister
 
 from ingeniamotion.exceptions import IMStatusWordError
 from ingeniamotion.enums import OperationMode, MonitoringSoCType, MonitoringSoCConfig
+from ingeniamotion.capture import map_register_address
 
 
 def __compare_signals(expected_signal, received_signal, fft_tol=0.05):
@@ -280,3 +283,34 @@ def test_get_frequency(
     new_divider = 2
     monitoring.set_frequency(new_divider)
     assert mc.capture.get_frequency(servo=alias) == max_frequency / new_divider
+
+
+@pytest.mark.parametrize(
+    "subnode, address, mapped_address_eth, mapped_address_can",
+    [
+        (1, 0x0010, 0x0010, 0x0010),
+        (2, 0x0020, 0x0820, 0x0020),
+        (3, 0x0030, 0x1030, 0x0030),
+    ],
+)
+def test_map_register_address(subnode, address, mapped_address_eth, mapped_address_can):
+    ethernet_param_dict = {
+        'subnode': subnode,
+        'address': address,
+        'dtype': REG_DTYPE.U16,
+        'access': REG_ACCESS.RW
+    }
+    canopen_param_dict = {
+        'subnode': subnode,
+        'idx': address,
+        'subidx': 0x00,
+        'dtype': REG_DTYPE.U16,
+        'access': REG_ACCESS.RW,
+        'identifier': '',
+        'units': '',
+        'cyclic': 'CONFIG'
+    }
+    register = EthernetRegister(**ethernet_param_dict)
+    assert mapped_address_eth == map_register_address(register)
+    register = CanopenRegister(**canopen_param_dict)
+    assert mapped_address_can == map_register_address(register)
